+++ app/src/main/java/afkt/project/model/item/ButtonValue.java
@@ -98,6 +98,7 @@ public class ButtonValue {
         lists.add(new ButtonValue(BTN_FASTJSON, "FastjsonUtils"));
         lists.add(new ButtonValue(BTN_ZXING, "ZXingQRCodeUtils"));
         lists.add(new ButtonValue(BTN_PICTURE_SELECTOR, "PictureSelectorUtils"));
+        lists.add(new ButtonValue(BTN_OKGO, "OkGoUtils"));
         return lists;
     }
 
@@ -375,6 +376,7 @@ public class ButtonValue {
         lists.add(new ButtonValue(BTN_SIGN_VIEW, "签名 View"));
         lists.add(new ButtonValue(BTN_LINE_VIEW, "换行监听 View"));
         lists.add(new ButtonValue(BTN_LIKE_VIEW, "自定义点赞效果 View"));
+        lists.add(new ButtonValue(BTN_CORNER_LABEL_VIEW, "自定义角标 View"));
         return lists;
     }
 
@@ -454,6 +456,9 @@ public class ButtonValue {
 
     // PictureSelectorUtils
     public static final int BTN_PICTURE_SELECTOR = MODULE_LIB + 6;
+
+    // OkGoUtils
+    public static final int BTN_OKGO = MODULE_LIB + 7;
 
     // ======
     // = UI =
@@ -719,6 +724,9 @@ public class ButtonValue {
 
     // 自定义点赞效果 View
     public static final int BTN_LIKE_VIEW = BTN_DEV_WIDGET + 700;
+
+    // 自定义角标 View
+    public static final int BTN_CORNER_LABEL_VIEW = BTN_DEV_WIDGET + 800;
 
     // ================================
     // = DevEnvironment 环境配置切换库 =