+++ lib/Widget/DevWidget/src/main/java/dev/widget/adjust/AdjustHeightGridView.java
@@ -1,6 +1,8 @@
 package dev.widget.adjust;
 
+import android.annotation.TargetApi;
 import android.content.Context;
+import android.os.Build;
 import android.util.AttributeSet;
 import android.widget.GridView;
 
@@ -20,6 +22,11 @@ public class AdjustHeightGridView extends GridView {
 
     public AdjustHeightGridView(Context context, AttributeSet attrs, int defStyleAttr) {
         super(context, attrs, defStyleAttr);
+    }
+
+    @TargetApi(Build.VERSION_CODES.LOLLIPOP)
+    public AdjustHeightGridView(Context context, AttributeSet attrs, int defStyleAttr, int defStyleRes) {
+        super(context, attrs, defStyleAttr, defStyleRes);
     }
 
     @Override