+++ lib/Widget/DevWidget/src/main/java/dev/widget/custom/CustomGallery.java
@@ -1,7 +1,9 @@
 package dev.widget.custom;
 
+import android.annotation.TargetApi;
 import android.content.Context;
 import android.content.res.TypedArray;
+import android.os.Build;
 import android.util.AttributeSet;
 import android.view.KeyEvent;
 import android.view.MotionEvent;
@@ -22,11 +24,11 @@ import dev.widget.R;
 public class CustomGallery extends Gallery {
 
     // 是否允许滑动
-    private boolean mIsSlide = true;
+    private boolean mIsSlide   = true;
     // 最大显示宽度
-    private int mMaxWidth = WidgetUtils.DEF_VALUE;
+    private int     mMaxWidth  = WidgetUtils.DEF_VALUE;
     // 最大显示高度
-    private int mMaxHeight = WidgetUtils.DEF_VALUE;
+    private int     mMaxHeight = WidgetUtils.DEF_VALUE;
 
     public CustomGallery(Context context) {
         super(context);
@@ -39,6 +41,12 @@ public class CustomGallery extends Gallery {
 
     public CustomGallery(Context context, AttributeSet attrs, int defStyleAttr) {
         super(context, attrs, defStyleAttr);
+        initAttrs(context, attrs);
+    }
+
+    @TargetApi(Build.VERSION_CODES.LOLLIPOP)
+    public CustomGallery(Context context, AttributeSet attrs, int defStyleAttr, int defStyleRes) {
+        super(context, attrs, defStyleAttr, defStyleRes);
         initAttrs(context, attrs);
     }