+++ lib/Widget/DevWidget/src/main/java/dev/widget/custom/CustomScrollView.java
@@ -1,8 +1,10 @@
 package dev.widget.custom;
 
+import android.annotation.TargetApi;
 import android.content.Context;
 import android.content.res.TypedArray;
 import android.graphics.Rect;
+import android.os.Build;
 import android.util.AttributeSet;
 import android.view.MotionEvent;
 import android.widget.ScrollView;
@@ -24,11 +26,11 @@ import dev.widget.R;
 public class CustomScrollView extends ScrollView {
 
     // 是否允许滑动
-    private boolean mIsSlide = true;
+    private boolean        mIsSlide        = true;
     // 最大显示宽度
-    private int mMaxWidth = WidgetUtils.DEF_VALUE;
+    private int            mMaxWidth       = WidgetUtils.DEF_VALUE;
     // 最大显示高度
-    private int mMaxHeight = WidgetUtils.DEF_VALUE;
+    private int            mMaxHeight      = WidgetUtils.DEF_VALUE;
     // 滑动监听回调
     private ScrollCallBack mScrollCallBack = null;
 
@@ -43,6 +45,12 @@ public class CustomScrollView extends ScrollView {
 
     public CustomScrollView(Context context, AttributeSet attrs, int defStyleAttr) {
         super(context, attrs, defStyleAttr);
+        initAttrs(context, attrs);
+    }
+
+    @TargetApi(Build.VERSION_CODES.LOLLIPOP)
+    public CustomScrollView(Context context, AttributeSet attrs, int defStyleAttr, int defStyleRes) {
+        super(context, attrs, defStyleAttr, defStyleRes);
         initAttrs(context, attrs);
     }