+++ lib/Widget/DevWidget/src/main/java/dev/widget/ui/FlowLikeView.java
@@ -6,17 +6,18 @@ import android.animation.AnimatorSet;
 import android.animation.ObjectAnimator;
 import android.animation.TypeEvaluator;
 import android.animation.ValueAnimator;
+import android.annotation.TargetApi;
 import android.content.Context;
 import android.content.res.TypedArray;
 import android.graphics.PointF;
 import android.graphics.drawable.Drawable;
+import android.os.Build;
+import android.support.annotation.DrawableRes;
 import android.util.AttributeSet;
 import android.view.Gravity;
 import android.view.View;
 import android.widget.FrameLayout;
 import android.widget.ImageView;
-
-import androidx.annotation.DrawableRes;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -47,13 +48,13 @@ public class FlowLikeView extends FrameLayout {
     // View 宽度、高度
     private int mViewWidth, mViewHeight;
     // 添加动画 View Layout 参数
-    private LayoutParams mLayoutParams;
+    private LayoutParams   mLayoutParams;
     // 用于产生随机数
-    private Random mRandom;
+    private Random         mRandom;
     // Icon 集合
     private List<Drawable> mDrawables;
     // 点赞 Icon 宽高
-    private int mIconWidth, mIconHeight;
+    private int            mIconWidth, mIconHeight;
     // 点赞动画执行时间
     private long mAnimDuration = 2000l;
 
@@ -69,6 +70,12 @@ public class FlowLikeView extends FrameLayout {
 
     public FlowLikeView(Context context, AttributeSet attrs, int defStyleAttr) {
         super(context, attrs, defStyleAttr);
+        initAttrs(context, attrs);
+    }
+
+    @TargetApi(Build.VERSION_CODES.LOLLIPOP)
+    public FlowLikeView(Context context, AttributeSet attrs, int defStyleAttr, int defStyleRes) {
+        super(context, attrs, defStyleAttr, defStyleRes);
         initAttrs(context, attrs);
     }
 
@@ -155,7 +162,7 @@ public class FlowLikeView extends FrameLayout {
         @Override
         public void onAnimationUpdate(ValueAnimator animation) {
             if (target != null) {
-                // 获取当前动画运行的状态值, 使得动x画作用对象沿着曲线 ( 涉及贝塞儿曲线 ) 运动
+                // 获取当前动画运行的状态值, 使得动 x 画作用对象沿着曲线 ( 涉及贝塞儿曲线 ) 运动
                 PointF pointF = (PointF) animation.getAnimatedValue();
                 target.setX(pointF.x);
                 target.setY(pointF.y);
@@ -182,9 +189,9 @@ public class FlowLikeView extends FrameLayout {
             PointF resultPointF = new PointF();
             // 二阶贝塞儿曲线
             resultPointF.x = (float) Math.pow(leftTime, 2) * startValue.x + 2 * fraction * leftTime * ctrlPointF.x
-                + ((float) Math.pow(fraction, 2)) * endValue.x;
+                    + ((float) Math.pow(fraction, 2)) * endValue.x;
             resultPointF.y = (float) Math.pow(leftTime, 2) * startValue.y + 2 * fraction * leftTime * ctrlPointF.y
-                + ((float) Math.pow(fraction, 2)) * endValue.y;
+                    + ((float) Math.pow(fraction, 2)) * endValue.y;
             return resultPointF;
         }
     }
@@ -209,8 +216,8 @@ public class FlowLikeView extends FrameLayout {
     private ValueAnimator generateCurveAnimation(View target) {
         CurveEvaluator evaluator = new CurveEvaluator(generateCTRLPointF(1));
         ValueAnimator valueAnimator = ValueAnimator.ofObject(evaluator,
-            new PointF((mViewWidth - mIconWidth) / 2, mViewHeight - mChildViewHeight - mIconHeight),
-            new PointF((mViewWidth) / 2 + (mRandom.nextBoolean() ? 1 : -1) * mRandom.nextInt(100), 0)
+                new PointF((mViewWidth - mIconWidth) / 2, mViewHeight - mChildViewHeight - mIconHeight),
+                new PointF((mViewWidth) / 2 + (mRandom.nextBoolean() ? 1 : -1) * mRandom.nextInt(100), 0)
         );
         valueAnimator.setDuration(mAnimDuration);
         valueAnimator.addUpdateListener(new CurveUpdateLister(target));